backend/requirement.txt
asyncpg
httpx
fastapi
oauthlib
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
requests-oauthlib
sqlalchemy
uvicorn